fix(Vercel): separate authOptions file
